+++ wetterdienst/provider/geosphere/observation/__init__.py
@@ -1,0 +1,6 @@
+from wetterdienst.provider.geosphere.observation.api import (
+    GeosphereObservationDataset,
+    GeosphereObservationParameter,
+    GeosphereObservationRequest,
+    GeosphereObservationResolution,
+)